streamlit-project/pyproject.toml
[project]
name = "streamlit-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt>=5.0.0",
    "fpdf2>=2.8.7",
    "google-genai>=1.68.0",
    "openpyxl>=3.1.5",
    "streamlit>=1.55.0",
]
